Protected the deleting route

+++ app/Http/Controllers/MealController.php
@@ -3,9 +3,11 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Symfony\Component\HttpFoundation\Response;
 use App\Http\Requests\StoreMealRequest;
 use App\Http\Requests\UpdateMealRequest;
 use App\Models\Meal;
+use Gate;
 
 class MealController extends Controller
 {
@@ -44,6 +46,8 @@ class MealController extends Controller
      */
     public function destroy(Meal $meal)
     {
+        abort_if(Gate::denies('meal_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        
         $meal->delete();
     }
 }
